requirements.txt
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
build==1.1.1
bz2file==0.98
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
dacite==1.8.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
docutils==0.20.1
doqu==0.28.2
et-xmlfile==1.1.0
executing==2.0.1
fastjsonschema==2.19.1
Flask==3.0.2
fonttools==4.49.0
fqdn==1.5.1
gitdb==4.0.11
GitPython==3.1.41
h11==0.14.0
htmlmin==0.1.12
httpcore==1.0.4
httpx==0.27.0
idna==3.6
ImageHash==4.3.1
importlib-metadata==6.11.0
ipykernel==6.29.3
ipython==8.12.3
ipywidgets==8.1.2
isoduration==20.11.0
itsdangerous==2.1.2
jaraco.classes==3.3.1
jedi==0.19.1
Jinja2==3.1.3
joblib==1.1.1
json5==0.9.22
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter-console==6.6.3
jupyter-contrib-core==0.4.2
jupyter-contrib-nbextensions==0.7.0
jupyter-events==0.9.0
jupyter-highlight-selected-word==0.2.0
jupyter-lsp==2.2.4
jupyter-nbextensions-configurator==0.6.3
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.13.0
jupyter_server_terminals==0.5.2
jupyterlab==4.1.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
jupyterlab_widgets==3.0.10
keyring==24.3.1
kiwisolver==1.4.5
llvmlite==0.41.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
missingno==0.5.2
mistune==3.0.2
mlxtend==0.23.1
more-itertools==10.2.0
multimethod==1.11.1
myfiglet==0.0.9
nbclient==0.9.0
nbconvert==7.16.2
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
nh3==0.2.15
nltk==3.8.1
notebook_shim==0.2.4
numba==0.58.1
numpy==1.24.2
openpyxl==3.1.2
overrides==7.7.0
packaging==23.2
pandas==1.5.3
pandas-profiling==3.2.0
pandocfilters==1.5.1
parso==0.8.3
patsy==0.5.6
phik==0.12.4
pickleshare==0.7.5
Pillow==9.5.0
pipreqs==0.5.0
pkginfo==1.10.0
platformdirs==4.2.0
plotly==5.19.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.8
pure-eval==0.2.2
py4math==0.0.3
pyarrow==14.0.2
pycparser==2.21
pydantic==2.6.1
pydantic-settings==2.2.1
pydantic_core==2.16.2
pydeck==0.8.1b0
Pygments==2.17.2
Pympler==1.0.1
pyparsing==3.1.1
pyproject_hooks==1.0.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.5.0
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
readme_renderer==43.0
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.17.1
scikit-learn==1.2.2
scipy==1.11.4
seaborn==0.12.2
Send2Trash==1.8.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
statsmodels==0.14.1
streamlit==1.24.1
streamlit-option-menu==0.3.6
streamlit-pandas-profiling==0.1.3
tangled-up-in-unicode==0.2.0
tenacity==8.2.3
terminado==0.18.0
threadpoolctl==3.2.0
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
twine==5.0.0
typeguard==4.1.5
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==4.3.1
uri-template==1.3.0
urllib3==2.1.0
validators==0.22.0
visions==0.7.5
watchdog==3.0.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
widgetsnbextension==4.0.10
wordcloud==1.9.3
yarg==0.1.9
ydata-profiling==4.6.4
zipp==3.17.0
